refactor: rename package structure for payment repository and tests
